refector: 랜덤한 우승 로또 생성 로직을 LottoGenerator라는 새로운 클래스로 분리
