+++ docs/README.md
@@ -66,7 +66,7 @@
 ### `LottoManager` 클래스에 몰려있는 주요 로직 분리
 
 - [x] 로또 간의 숫자 비교 및 보너스 숫자 포함 여부 판단 로직을 `Lotto` 클래스로 분리한다.
-- [ ] 랜덤한 우승 로또 생성 로직을 `LottoGenerator`라는 새로운 클래스로 분리한다.
+- [x] 랜덤한 우승 로또 생성 로직을 `LottoGenerator`라는 새로운 클래스로 분리한다.
 
 ### `LottoManager` 및 `Player` 클래스의 역할 재정의
